refactor postState check add remote postState check

+++ retesteth/RPCSession.h
@@ -50,7 +50,7 @@ public:
     std::string test_getLogHash(std::string const& _txHash);
 	void test_setChainParams(std::string const& _config);
 	void test_rewindToBlock(size_t _blockNr);
-	void test_modifyTimestamp(size_t _timestamp);
+    void test_modifyTimestamp(unsigned long long _timestamp);
     void test_mineBlocks(int _number);
     void test_importRawBlock(std::string const& _blockRLP);
